pyproject.toml
[tool.poetry]
name = "project self-study"
version = "0.1.0"
description = ""
authors = ["AlexandraAvtaykina <justalexandra@mail.ru>"]
readme = "README.md"
packages = [{include = "project self_study"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.2"
ipython = "^8.21.0"
pillow = "^10.2.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
flake8 = "^7.0.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
